+++ backend/src/main/java/unischedule/exception/GlobalExceptionHandler.java
@@ -4,6 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import unischedule.exception.dto.EntityAlreadyExistsException;
 import unischedule.exception.dto.ErrorResponseDto;
 
 @RestControllerAdvice
@@ -11,7 +12,7 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(EntityNotFoundException.class)
     public ResponseEntity<ErrorResponseDto> handleEntityNotFound(EntityNotFoundException ex) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(new ErrorResponseDto(ex.getMessage()));
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ErrorResponseDto.of(ex.getMessage()));
     }
 
     @ExceptionHandler(InvalidInputException.class)
@@ -21,7 +22,12 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(ExternalApiException.class)
     public ResponseEntity<ErrorResponseDto> handleEntityNotFound(ExternalApiException ex) {
-        return ResponseEntity.internalServerError().body(new ErrorResponseDto(ex.getMessage()));
+        return ResponseEntity.internalServerError().body(ErrorResponseDto.of(ex.getMessage()));
+    }
+
+    @ExceptionHandler(EntityAlreadyExistsException.class)
+    public ResponseEntity<ErrorResponseDto> handleEntityAlreadyExists(EntityAlreadyExistsException ex) {
+        return ResponseEntity.badRequest().body(ErrorResponseDto.of(ex.getMessage()));
     }
 }
