+++ backend/src/main/java/unischedule/exception/dto/EntityAlreadyExistsException.java
@@ -1,0 +1,8 @@
+package unischedule.exception.dto;
+
+public class EntityAlreadyExistsException extends RuntimeException {
+
+    public EntityAlreadyExistsException(String message) {
+        super(message);
+    }
+}
