+++ backend/src/main/java/unischedule/member/controller/MemberApiController.java
@@ -6,7 +6,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
@@ -29,10 +28,6 @@ public class MemberApiController {
 
     @PostMapping("/signup")
     public ResponseEntity<String> signup(@Valid @RequestBody MemberRegistrationDto requestDto) {
-        if (memberService.isMemberExists(requestDto.email())) {
-            return ResponseEntity.badRequest().body("이미 사용중인 이메일입니다.");
-        }
-
         memberService.registerMember(requestDto);
 
         return ResponseEntity.status(HttpStatus.CREATED).body("회원가입이 완료되었습니다.");
