+++ backend/src/test/java/unischedule/member/controller/MemberApiControllerTest.java
@@ -9,22 +9,23 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.core.Authentication;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 import unischedule.auth.jwt.JwtTokenProvider;
+import unischedule.exception.dto.EntityAlreadyExistsException;
 import unischedule.member.dto.LoginRequestDto;
 import unischedule.member.dto.MemberRegistrationDto;
 import unischedule.member.service.MemberService;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.doNothing;
+import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -52,7 +53,6 @@ class MemberApiControllerTest {
     void signup() throws Exception {
         // given
         MemberRegistrationDto registrationDto = new MemberRegistrationDto("test12@gmail.com", "nickname", "p1a2s3s4word");
-        given(memberService.isMemberExists(registrationDto.email())).willReturn(false);
         doNothing().when(memberService).registerMember(any(MemberRegistrationDto.class));
 
         // when & then
@@ -68,14 +68,17 @@ class MemberApiControllerTest {
     void signupDuplicated() throws Exception {
         // given
         MemberRegistrationDto registrationDto = new MemberRegistrationDto("test12@gmail.com", "nickname", "p1a2s3s4word");
-        given(memberService.isMemberExists(registrationDto.email())).willReturn(true);
+
+        doThrow(new EntityAlreadyExistsException("이미 사용중인 이메일입니다."))
+                .when(memberService)
+                .registerMember(any(MemberRegistrationDto.class));
 
         // when & then
         mockMvc.perform(post("/api/members/signup")
                 .contentType(MediaType.APPLICATION_JSON)
                 .content(objectMapper.writeValueAsString(registrationDto)))
                 .andExpect(status().isBadRequest())
-                .andExpect(jsonPath("$").value("이미 사용중인 이메일입니다."));
+                .andExpect(jsonPath("$.message").value("이미 사용중인 이메일입니다."));
     }
 
     @Test
